Broke controllers and services

+++ codebase/backend/core/src/controllers/user-controller.ts
@@ -3,7 +3,7 @@ import { Request, Response } from "express";
 import { UserService } from "../services/user-service";
 
 import initializeLogger from "../logger";
-import { userInfo } from "os";
+import { IUserSearchOptions, Role } from "../types";
 
 const log = initializeLogger(__filename.split("\\").pop() || "");
 
@@ -11,7 +11,6 @@ const getUser = async (req: Request, res: Response) => {
 	try {
 		log.info("Attempting to find user");
 		const { id }: { id: string } = req.params as { id: string };
-		log.info(`Id is : ${id}`);
 		const user = await UserService.getUser(id);
 		log.info("Successfully found user");
 		return res.status(HttpStatusCode.Ok).send({ ...user });
@@ -26,26 +25,17 @@ const getUser = async (req: Request, res: Response) => {
 const searchUsers = async (req: Request, res: Response) => {
 	try {
 		log.info("Attempting to search users");
-		const {
-			pageNum,
-			pageSize,
-			sortCol,
-			sortDir,
-			search,
-		}: {
-			pageNum?: number;
-			pageSize?: number;
-			sortCol?: string;
-			sortDir?: "asc" | "desc";
-			search?: string;
-		} = req.body;
-		const users = await UserService.searchUsers({
-			pageNum,
-			pageSize,
-			sortCol,
-			sortDir,
-			search,
-		});
+		const searchOptions: IUserSearchOptions = req.body;
+		const authorizedUserRoles = req.headers["user-roles"] as Role[] | undefined;
+		if (
+			authorizedUserRoles?.includes(Role.SELLER) ||
+			authorizedUserRoles?.includes(Role.BUYER)
+		) {
+			return res
+				.status(HttpStatusCode.Unauthorized)
+				.send("Users cannot be searched by Buyers or Sellers");
+		}
+		const users = await UserService.searchUsers(searchOptions);
 		log.info("Created list of users");
 		return res.status(HttpStatusCode.Ok).send(users);
 	} catch (err) {
@@ -71,6 +61,14 @@ const editUser = async (req: Request, res: Response) => {
 			profileImageUrl?: string;
 		} = req.body;
 		const { id } = req.params as { id: string };
+		const authorizedUserId = req.headers["user-id"] as string | undefined;
+		const authorizedUserRoles = req.headers["user-roles"] as Role[] | undefined;
+
+		if (authorizedUserId !== id && !authorizedUserRoles?.includes(Role.ADMIN)) {
+			return res
+				.status(HttpStatusCode.Unauthorized)
+				.send("Unauthorized to edit user");
+		}
 
 		const user = await UserService.editUser({
 			id,
@@ -93,6 +91,13 @@ const deleteUser = async (req: Request, res: Response) => {
 	try {
 		log.info("Attempting to delete user");
 		const { id } = req.params as { id: string };
+		const authorizedUserId = req.headers["user-id"] as string | undefined;
+		const authorizedUserRoles = req.headers["user-roles"] as Role[] | undefined;
+		if (authorizedUserId !== id && !authorizedUserRoles?.includes(Role.ADMIN)) {
+			return res
+				.status(HttpStatusCode.Unauthorized)
+				.send("Unauthorized to delete user");
+		}
 		await UserService.deleteUser(id);
 		log.info("Deleted user");
 		return res.status(HttpStatusCode.Ok).send();
